Login Page: elements Fixes #11
